feature(model): adjustments to the DECREE and creation of queries for total comments and reactions

+++ emi/src/main/java/com/conexemi/emi/services/CommentsService.java
@@ -5,6 +5,8 @@ import com.conexemi.emi.Exceptions.ResourceNotFoundException;
 import com.conexemi.emi.Mapper.CommentsMapper;
 import com.conexemi.emi.model.Comments;
 import com.conexemi.emi.repositories.CommentsRepository;
+import com.conexemi.emi.repositories.EntrepreneurshipRepository;
+import com.conexemi.emi.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -19,35 +21,44 @@ public class CommentsService {
     private CommentsRepository commentsRepository;
 
     @Autowired
-    private CommentsMapper commentsMapper;
+    private EntrepreneurshipRepository entrepreneurshipRepository;
+
+    @Autowired
+    private UserRepository userRepository;
+
+    @Autowired
+    public CommentsService(CommentsRepository commentsRepository) {
+        this.commentsRepository = commentsRepository;
+    }
 
 
     public CommentsDTO createComment(CommentsDTO commentDTO) {
-        Comments comment = commentsMapper.toEntity(commentDTO);
+        Comments comment = CommentsMapper.toEntity(commentDTO, entrepreneurshipRepository, userRepository);
         Comments savedComment = commentsRepository.save(comment);
-        return commentsMapper.toDTO(savedComment);
+        return CommentsMapper.toDTO(savedComment, commentsRepository);
     }
 
     public Optional<CommentsDTO> getCommentById(Integer idComment) {
-        Optional<Comments> comment = Optional.ofNullable(commentsRepository.findById(idComment)
-                .orElseThrow(() -> new ResourceNotFoundException("Comments with ID " + idComment + " not found")));
-        return comment.map(CommentsMapper::toDTO);
+        Comments comment = commentsRepository.findById(idComment)
+                .orElseThrow(() -> new ResourceNotFoundException("Comment with ID " + idComment + " not found"));
+        return Optional.of(CommentsMapper.toDTO(comment, commentsRepository));
     }
 
     public List<CommentsDTO> getAllComments() {
         List<Comments> comments = commentsRepository.findAll();
-        return comments.stream().map(CommentsMapper::toDTO).collect(Collectors.toList());
+        return comments.stream().map(comment -> CommentsMapper.toDTO(comment, commentsRepository)).collect(Collectors.toList());
     }
 
     public void deleteCommentById(Integer idComment) {
+        if (!commentsRepository.existsById(idComment)) {
+            throw new ResourceNotFoundException("Comment with ID " + idComment + " not found");
+        }
         commentsRepository.deleteById(idComment);
     }
 
     public List<CommentsDTO> getCommentsByEntrepreneurship(Integer idEntrepreneurship) {
         List<Comments> comentsList = commentsRepository.getCommentsByEntrepreneurship(idEntrepreneurship);
-        return comentsList.stream()
-                .map(comments -> CommentsMapper.toDTO(comments))
-                .collect(Collectors.toList());
+        return comentsList.stream().map(comment -> CommentsMapper.toDTO(comment, commentsRepository)).collect(Collectors.toList());
     }
